+++ tests/testthat/test-Assigned.R
@@ -19,7 +19,7 @@ test_that("Poisson Assigned Events, check results", {
   keep_constant <- c(0, 0, 0, 0)
 
   control <- list(
-    "ncores" = 2, "lr" = 0.75, "maxiter" = 1, "halfmax" = 5, "epsilon" = 1e-3,
+    "ncores" = 1, "lr" = 0.75, "maxiter" = 1, "halfmax" = 5, "epsilon" = 1e-3,
     "deriv_epsilon" = 1e-3, "step_max" = 1.0, "change_all" = TRUE,
     "thres_step_max" = 100.0, "verbose" = 0
   )
@@ -57,8 +57,8 @@ test_that("Poisson Assigned Events, check results strata", {
   keep_constant <- c(0, 0, 0)
 
   control <- list(
-    "ncores" = 2, "lr" = 0.75, "maxiter" = 1, "halfmax" = 5, "epsilon" = 1e-3,
-    "deriv_epsilon" = 1e-3, "step_max" = 1.0, "change_all" = TRUE,
+    "ncores" = 1, "lr" = 0.75, "maxiter" = 1, "halfmax" = 5, "epsilon" = 1e-3,
+    "deriv_epsilon" = 1e-3, "step_max" = 0.2, "change_all" = TRUE,
     "thres_step_max" = 100.0, "verbose" = 0
   )
   #
@@ -71,10 +71,15 @@ test_that("Poisson Assigned Events, check results strata", {
   if (!isTRUE(as.logical(Sys.getenv("NOT_CRAN", "false")))) {
     skip("Cran Skip")
   }
+  control <- list(
+    "ncores" = 2, "lr" = 0.75, "maxiter" = 1, "halfmax" = 5, "epsilon" = 1e-3,
+    "deriv_epsilon" = 1e-3, "step_max" = 0.2, "change_all" = TRUE,
+    "thres_step_max" = 100.0, "verbose" = 0
+  )
   for (i in 1:9) {
     a_n <- 2 * runif(3) - 1
     poisres$beta_0 <- a_n
-    e <- EventAssignment(poisres, df)
+    e <- EventAssignment(poisres, df, control = control)
 
     e0 <- e$predict
     e1 <- e$caused
@@ -83,7 +88,7 @@ test_that("Poisson Assigned Events, check results strata", {
     expect_equal(sum(e1[, 1:2]), sum(e1[, 3]), tolerance = 1e-2)
 
     poisres_strata$beta_0 <- a_n
-    e <- EventAssignment(poisres_strata, df)
+    e <- EventAssignment(poisres_strata, df, control = control)
 
     e0 <- e$predict
     e1 <- e$caused
@@ -92,7 +97,7 @@ test_that("Poisson Assigned Events, check results strata", {
     expect_equal(sum(e1[, 1:2]), sum(e1[, 3]), tolerance = 1e-2)
   }
 })
-#
+
 test_that("Poisson Assigned Events, combinations", {
   df <- data.table::data.table(
     "UserID" = c(112, 114, 213, 214, 115, 116, 117),
@@ -111,16 +116,16 @@ test_that("Poisson Assigned Events, combinations", {
 
   keep_constant <- c(0, 0, 0, 0)
   control <- list(
-    "ncores" = 2, "lr" = 0.75, "maxiter" = 1, "halfmax" = 5,
+    "ncores" = 1, "lr" = 0.75, "maxiter" = 1, "halfmax" = 5,
     "epsilon" = 1e-3, "deriv_epsilon" = 1e-3,
-    "step_max" = 1.0, "change_all" = TRUE, "thres_step_max" = 100.0,
+    "step_max" = 0.2, "change_all" = TRUE, "thres_step_max" = 100.0,
     "verbose" = 0, "ties" = "breslow"
   )
   poisres <- PoisRun(Pois(pyr, Cancer_Status) ~ loglinear(a, 0) + loglinear(b, 1) + loglinear(c, 2), df, a_n = a_n, control = control)
   df$Cancer_Status <- rep(0, nrow(df))
   expect_error(EventAssignment(poisres, df))
 })
-#
+
 test_that("Poisson Assigned Events bounds, check results", {
   df <- data.table::data.table(
     "UserID" = c(112, 114, 213, 214, 115, 116, 117),
@@ -137,8 +142,8 @@ test_that("Poisson Assigned Events bounds, check results", {
   a_n <- c(-0.75, 0.1, -0.05, -1.5)
   keep_constant <- c(0, 0, 0, 0)
   control <- list(
-    "ncores" = 2, "lr" = 0.75, "maxiter" = 100, "halfmax" = 5, "epsilon" = 1e-3,
-    "deriv_epsilon" = 1e-3, "step_max" = 1.0, "change_all" = TRUE,
+    "ncores" = 1, "lr" = 0.75, "maxiter" = 100, "halfmax" = 5, "epsilon" = 1e-3,
+    "deriv_epsilon" = 1e-3, "step_max" = 0.2, "change_all" = TRUE,
     "thres_step_max" = 100.0, "verbose" = 0
   )
   #
@@ -182,8 +187,8 @@ test_that("Poisson Assigned Events bounds single entry, check results", {
   a_n <- c(-0.75)
   keep_constant <- c(0)
   control <- list(
-    "ncores" = 2, "lr" = 0.75, "maxiter" = 100, "halfmax" = 5, "epsilon" = 1e-3,
-    "deriv_epsilon" = 1e-3, "step_max" = 1.0, "change_all" = TRUE,
+    "ncores" = 1, "lr" = 0.75, "maxiter" = 100, "halfmax" = 5, "epsilon" = 1e-3,
+    "deriv_epsilon" = 1e-3, "step_max" = 0.2, "change_all" = TRUE,
     "thres_step_max" = 100.0, "verbose" = 0
   )
   #