+++ NEWS.md
@@ -156,3 +156,9 @@
 
 * Updated geometric-mixture code to apply defaults of 0.5 theta and only excess terms
 * Corrected normalization for intercept parameters
+
+# Colossus 1.4.6
+
+* Fixed bug in second derivative risk calculations, wrong correction to multi-term risk values.
+* Improved results printing. Basic regression results only print term number column when multiple terms used. Likelihood boundary results now print the limit and score when negative limit hit.
+* ncores option used for Colossus calculations now also applied to data.table operations. Previous number reset after calculations, may need to be manually reset if the regression hits an error.